Front validator added, Mapper test added and some methods fixed

+++ computerdatabase/src/main/java/model/Computer.java
@@ -79,7 +79,7 @@ public class Computer extends Entity{
 	
 	@Override
 	public String toString() {
-		return "'" +this.getId().longValue() + "','" +
+		return "'" +this.getId()+ "','" +
 				this.getName() + "','" +
 				this.getIntroduced() + "','" +
 				this.getDiscontinued() + "','" +
